reconfiguring gui stuff
